Factor out calling the plotter.
